# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  auth.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user: {}
    type: object
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  auth.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  auth.RegisterRequest:
    properties:
      email:
        type: string
      full_name:
        type: string
      password:
        minLength: 6
        type: string
      phone:
        type: string
      role:
        type: string
    required:
    - email
    - full_name
    - password
    type: object
  handlers.Notification:
    properties:
      createdAt:
        type: string
      id:
        type: string
      isRead:
        type: boolean
      message:
        type: string
      type:
        description: '"NEW_REQUEST", "NEW_RETAILER"'
        type: string
    type: object
  handlers.ServiceApplication:
    properties:
      amount:
        type: number
      applicationId:
        type: string
      createdDate:
        type: string
      formData:
        additionalProperties: true
        type: object
      retailerId:
        type: string
      retailerName:
        type: string
      serviceId:
        type: string
      serviceName:
        type: string
      status:
        type: string
    type: object
  wallet.GatewayRechargeRequestPayload:
    properties:
      amount:
        type: number
      customer_email:
        type: string
      customer_mobile:
        type: string
      redirect_url:
        type: string
    required:
    - amount
    - customer_email
    - customer_mobile
    - redirect_url
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the backend API for the E-Service platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: E-Service API
  version: "1.0"
paths:
  /admin/notifications:
    get:
      description: Get a list of notifications for the Admin Panel
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.Notification'
            type: array
      summary: Get Notifications
      tags:
      - Admin
  /admin/requests:
    get:
      description: Fetch all service requests simultaneously for Admin Panel
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.ServiceApplication'
            type: array
      summary: Get Admin Service Requests
      tags:
      - Admin
  /admin/retailer:
    post:
      consumes:
      - application/json
      description: Add a new Retailer or Distributor and trigger a notification
      parameters:
      - description: User Data
        in: body
        name: request
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Add Retailer or Distributor
      tags:
      - Admin
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user with email and password to receive tokens.
      parameters:
      - description: Login Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout the user by invalidating their refresh token.
      parameters:
      - description: Logout Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout user
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh the JWT access token using a refresh token.
      parameters:
      - description: Refresh Token Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh Access Token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with email, password, full name, phone, and
        role.
      parameters:
      - description: Registration Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
  /services/request:
    post:
      consumes:
      - application/json
      description: Retailer or Distributor submits a new service request application
      parameters:
      - description: Service Application Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ServiceApplication'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Submit a Service Request
      tags:
      - Services
  /wallet/recharge/gateway:
    post:
      consumes:
      - application/json
      description: Creates a new Mugavi payment gateway order and returns the payment
        URL and QR code.
      parameters:
      - description: Gateway Recharge Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/wallet.GatewayRechargeRequestPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Initiate Payment Gateway Recharge
      tags:
      - Wallet
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"